Rework image and file I/O
