+++ Source/Makefile
@@ -44,10 +44,10 @@ SOURCES = \
     astc_error_metrics.cpp \
     astc_find_best_partitioning.cpp \
     astc_ideal_endpoints_and_weights.cpp \
+    astc_image.cpp \
     astc_image_load_store.cpp \
     astc_integer_sequence.cpp \
     astc_kmeans_partitioning.cpp \
-    astc_ktx_dds.cpp \
     astc_main.cpp \
     astc_mathlib.cpp \
     astc_mathlib_softfloat.cpp \
@@ -57,7 +57,6 @@ SOURCES = \
     astc_platform_isa_detection.cpp \
     astc_platform_dependents.cpp \
     astc_quantization.cpp \
-    astc_stb_tga.cpp \
     astc_symbolic_physical.cpp \
     astc_toplevel.cpp \
     astc_toplevel_help.cpp \
